Add file existence checker

+++ src/helpers.js
@@ -5,7 +5,7 @@
  */
 
 import { mkdir, writeFile } from 'node:fs/promises';
-import { getJSONConfigFile } from './utils.js';
+import { checkIfFileExists, getJSONConfigFile } from './utils.js';
 import {
   getComponentTemplate,
   getConfigTemplate,
@@ -20,27 +20,35 @@ const defaults = {
 };
 
 export async function createBoilerplate(componentName, dir, lang, styling) {
-  const componentFileExtensions = {
-    js: 'jsx',
-    ts: 'tsx',
-  };
-  const componentTemplate = await getComponentTemplate(
-    componentName,
-    lang,
-    styling
-  );
-  const indexTemplate = await getIndexTemplate(componentName);
+  try {
+    const componentFileExtensions = {
+      js: 'jsx',
+      ts: 'tsx',
+    };
+    const componentTemplate = await getComponentTemplate(
+      componentName,
+      lang,
+      styling
+    );
+    const indexTemplate = await getIndexTemplate(componentName);
 
-  if (!componentTemplate || !indexTemplate) {
-    throw new Error('Template code could not be found');
+    if (!componentTemplate || !indexTemplate) {
+      throw new Error('Template code could not be found.');
+    }
+
+    if (checkIfFileExists(dir)) {
+      throw new Error('This component already exists!');
+    }
+
+    await mkdir(dir, { recursive: true });
+    await writeFile(
+      `${dir}/${componentName}.${componentFileExtensions[lang]}`,
+      componentTemplate
+    );
+    await writeFile(`${dir}/index.${lang}`, indexTemplate);
+  } catch (e) {
+    console.error(e.message);
   }
-
-  await mkdir(dir, { recursive: true });
-  await writeFile(
-    `${dir}/${componentName}.${componentFileExtensions[lang]}`,
-    componentTemplate
-  );
-  await writeFile(`${dir}/index.${lang}`, indexTemplate);
 }
 
 export async function createConfig() {
@@ -60,11 +68,14 @@ export async function createConfig() {
 }
 
 export async function getOptions(options) {
-  const configFile = await getJSONConfigFile(configFileName);
-
-  return {
-    dir: options.dir || configFile?.dir || defaults.dir,
-    lang: options.lang || configFile?.lang || defaults.lang,
-    styling: options.styling || configFile?.lang || defaults.styling,
-  };
+  try {
+    const configFile = await getJSONConfigFile(configFileName);
+    return {
+      dir: options.dir || configFile?.dir || defaults.dir,
+      lang: options.lang || configFile?.lang || defaults.lang,
+      styling: options.styling || configFile?.lang || defaults.styling,
+    };
+  } catch (e) {
+    console.error(e);
+  }
 }
